Update tasks overview
