--- .github/workflows/php-test-develop.yml
@@ -1,13 +1,0 @@
-name: PHP Codeception Tests - develop
-
-on:
-  push:
-  schedule:
-    - cron: "0 0 * * 0"
-
-jobs:
-  tests:
-    uses: humhub/actions/.github/workflows/module-tests-develop.yml@main
-    with:
-      module-id: tasks
-      use-rest-module: true
