--- .github/workflows/php-test-next.yml
@@ -1,13 +1,0 @@
-name: PHP Codeception Tests - next
-
-on:
-  push:
-  schedule:
-    - cron: "0 0 * * 0"
-
-jobs:
-  tests:
-    uses: humhub/actions/.github/workflows/module-tests-next.yml@main
-    with:
-      module-id: tasks
-      use-rest-module: true
