+++ widgets/views/taskInfoBox.php
@@ -1,7 +1,7 @@
 <?php
 /* @var $this \humhub\modules\ui\view\components\View */
 /* @var $title string */
-/* @var $value string */
+/* @var $value array|string */
 /* @var $icon string */
 /* @var $iconColor string */
 /* @var $cssClass string */
@@ -14,12 +14,11 @@ use humhub\modules\ui\icon\widgets\Icon;
 ?>
 
 <div class="task-info <?= $cssClass ?>">
-    <strong><?= Icon::get($icon)->color($iconColor)?> <?= Html::encode($title) ?>
-    </strong><br>
+    <strong><?= Icon::get($icon)->color($iconColor)?> <?= Html::encode($title) ?></strong>
     <span class="task-info-text <?= $textClass ?>">
         <?php foreach ($value as $val) : ?>
             <?php if (!empty($value)) : ?>
-                <small><?= $val ?></small><br>
+                <?= $val ?><br>
             <?php endif; ?>
         <?php endforeach; ?>
     </span>
